starting update days of planning + validate journey

+++ workmanager/lib/layouts/home/contract/planning/day/day_tile.dart
@@ -8,6 +8,8 @@ import 'package:work_manager/models/planning.dart';
 import 'package:work_manager/models/user.dart';
 import 'package:work_manager/services/databases/planningDao.dart';
 
+import 'modify_time_slots.dart';
+
 
 
 class DayTile extends StatefulWidget{
@@ -25,13 +27,24 @@ class DayTile extends StatefulWidget{
 class _DayTileState extends State<DayTile>{
 
 
+  Icon indicator= Icon(Icons.fiber_manual_record,color: Colors.amber,);
+
+
+  @override
+  void initState() {
+    super.initState();
+    if(widget.dayData.endDate.difference(DateTime.now()).inMilliseconds >0)
+      indicator= Icon(Icons.fiber_manual_record,color: Colors.green);
+    else
+      indicator= Icon(Icons.fiber_manual_record,color: Colors.amber);
+  }
 
   @override
   Widget build(BuildContext context) {
 
     // TODO: implement build
     final user = Provider.of<User>(context);
-    if (widget.contractData.employerId == user.uid) {  //  controle la provenance de la proposition
+    if (widget.contractData.employerId != user.uid) {  //  controle la provenance de la proposition
       return GestureDetector(
         child: Padding(
           padding: EdgeInsets.only(top: 5.0),
@@ -49,36 +62,35 @@ class _DayTileState extends State<DayTile>{
                     Row(
                       children: <Widget>[
                         SizedBox(width: 10,),
-                        Text("heure début: ${widget.dayData.startDate
+                        Text("début: ${widget.dayData.startDate
                             .toString()
-                            .split(" ")[0]}        heure fin: ${widget
+                            .split(" ")[0]}        fin: ${widget
                             .dayData.endDate.toString().split(
                             " ")[0]}"),
+                        SizedBox(width: 20,),
+                        indicator,
                       ],
                     ),
                     SizedBox(height: 15,),
                     Row(
                       children: <Widget>[
                         Container(
-                          padding: EdgeInsets.symmetric(horizontal: 15),
+                          padding: EdgeInsets.symmetric(horizontal: 8),
                           child: RaisedButton(
                             color: Colors.white,
                             child: Container(
                               padding: EdgeInsets.symmetric(
                                   vertical: 5, horizontal: 5),
                               child: Row(children: <Widget>[
-                                Text("consulter",
+                                Text("plage horaire",
                                     style: TextStyle(color: Colors.black87)),
                                 SizedBox(width: 5,),
-                                Icon(Icons.calendar_today,
+                                Icon(Icons.edit,
                                   color: Colors.black87,)
                               ]),
                             ),
                             onPressed: () {
-                              Navigator.push(
-                                context,
-                                MaterialPageRoute(builder: (context) => ContractHome(contract: widget.contractData,)),
-                              );
+                              showTimeSlotModificationDialog( context,widget.dayData);
                             },
                           ),
                         ),
@@ -98,14 +110,21 @@ class _DayTileState extends State<DayTile>{
                               ]),
                             ),
                             onPressed: () async {
-                              try {
-                                await PlanningDao().deleteDayOfPlanning(
-                                    widget.dayData).then((value) {
-                                  Alert().goodAlert(context, "journée supprimée", "la journée identifiée par ${widget.dayData.documentId} a bien été supprimé");
-                                });
-                              }catch(e){
-                                print(e);
-
+                              if(widget.dayData.startDate.difference(DateTime.now()).inMinutes > 0) {
+                                try {
+                                  await PlanningDao().deleteDayOfPlanning(
+                                      widget.dayData).then((value) {
+                                    Alert().goodAlert(
+                                        context, "journée supprimée",
+                                        "la journée identifiée par ${widget
+                                            .dayData
+                                            .documentId} a bien été supprimé");
+                                  });
+                                } catch (e) {
+                                  print(e);
+                                }
+                              }else{
+                                Alert().badAlert(context, "suppression impossible", "la journée a deja commencé, suppression impossible");
                               }
                             },
                           ),
@@ -153,29 +172,45 @@ class _DayTileState extends State<DayTile>{
                     Row(
                       children: <Widget>[
                         Container(
-                          padding: EdgeInsets.symmetric(horizontal: 15),
+                          padding: EdgeInsets.symmetric(horizontal: 25),
                           child: RaisedButton(
                             color: Colors.white,
                             child: Container(
                               padding: EdgeInsets.symmetric(
                                   vertical: 5, horizontal: 5),
                               child: Row(children: <Widget>[
-                                Text("valider ",
+                                Text("arrivée ",
                                     style: TextStyle(color: Colors.black87)),
                                 SizedBox(width: 5,),
-                                Icon(Icons.collections,
+                                Icon(Icons.directions_run,
                                   color: Colors.black87,)
                               ]),
                             ),
                             onPressed: () {
-                              Navigator.push(
-                                context,
-                                MaterialPageRoute(builder: (context) =>  ContractHome(contract: widget.contractData)),
-                              );
+
                             },
                           ),
                         ),
+                        Container(
+                          padding: EdgeInsets.symmetric(horizontal: 0),
+                          child: RaisedButton(
+                            color: Colors.white,
+                            child: Container(
+                              padding: EdgeInsets.symmetric(
+                                  vertical: 5, horizontal: 5),
+                              child: Row(children: <Widget>[
+                                Text("départ ",
+                                    style: TextStyle(color: Colors.black87)),
+                                SizedBox(width: 5,),
+                                Icon(Icons.check_circle_outline,
+                                  color: Colors.black87,)
+                              ]),
+                            ),
+                            onPressed: () {
 
+                            },
+                          ),
+                        ),
 
                       ],
                     ),
